moved compose to a separate module, added traditional-xml-ui module, added BaseFragment and some other setup for xml module
