+++ app-compose/build.gradle
@@ -28,7 +28,11 @@ android {
         }
     }
     buildFeatures {
-        viewBinding true
+        compose true
+    }
+    composeOptions {
+        kotlinCompilerExtensionVersion version_compose
+        kotlinCompilerVersion '1.5.10'
     }
     packagingOptions {
         resources {
@@ -43,6 +47,20 @@ dependencies {
     implementation "com.google.dagger:hilt-android:$version_hilt"
     kapt "com.google.dagger:hilt-android-compiler:$version_hilt"
 
+    implementation "androidx.compose.ui:ui:$version_compose"
+    implementation "androidx.compose.material:material:$version_compose"
+    implementation "androidx.compose.ui:ui-tooling-preview:$version_compose"
+    implementation "androidx.activity:activity-compose:$version_appcompat"
+
+    implementation "com.google.accompanist:accompanist-insets:$version_insets"
+    implementation "com.google.accompanist:accompanist-insets-ui:$version_insets"
+
+//    implementation "androidx.lifecycle:lifecycle-viewmodel-compose:$version_compose"
+//    implementation "androidx.navigation:navigation-compose:2.4.0-alpha06"
+
     androidTestApi 'androidx.test.ext:junit:1.1.3'
     androidTestApi "androidx.test.espresso:espresso-core:$version_espresso"
+    androidTestApi "androidx.compose.ui:ui-test-junit4:$version_compose"
+    androidTestImplementation "androidx.compose.ui:ui-test-junit4:$version_compose"
+    debugImplementation "androidx.compose.ui:ui-tooling:$version_compose"
 }
