+++ app-compose/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.tinkooladik.tvshows">
+    package="com.tinkooladik.tvshows.compose">
 
     <application
         android:name=".TvShowsApplication"
@@ -11,7 +11,7 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.TvShows">
         <activity
-            android:name=".screen.MainActivity"
+            android:name=".NavActivity"
             android:exported="true"
             android:label="@string/app_name"
             android:theme="@style/Theme.TvShows.NoActionBar">
