+++ app-compose/src/main/java/com/tinkooladik/tvshows/compose/TvShowsApplication.kt
@@ -1,0 +1,19 @@
+package com.tinkooladik.tvshows.compose
+
+import android.app.Application
+import dagger.hilt.android.HiltAndroidApp
+import timber.log.Timber
+import javax.inject.Inject
+
+@HiltAndroidApp
+class TvShowsApplication : Application() {
+
+    @Inject
+    lateinit var tree: Timber.Tree
+
+    override fun onCreate() {
+        super.onCreate()
+
+        Timber.plant(tree)
+    }
+}
