+++ app/src/main/java/com/tinkooladik/tvshows/base/BaseFragment.kt
@@ -1,0 +1,24 @@
+package com.tinkooladik.tvshows.base
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+import androidx.viewbinding.ViewBinding
+
+abstract class BaseFragment<B : ViewBinding> : Fragment() {
+
+    protected lateinit var binding: B
+
+    override fun onCreateView(
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        binding = inflateView()
+        return binding.root
+    }
+
+    abstract fun inflateView(): B
+}
