+++ app/src/main/java/com/tinkooladik/tvshows/base/BaseViewModel.kt
@@ -3,14 +3,12 @@ package com.tinkooladik.tvshows.base
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import kotlinx.coroutines.channels.Channel
-import kotlinx.coroutines.flow.MutableStateFlow
-import kotlinx.coroutines.flow.asStateFlow
-import kotlinx.coroutines.flow.receiveAsFlow
+import kotlinx.coroutines.flow.*
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.sync.Mutex
 import kotlinx.coroutines.sync.withLock
 
-abstract class BaseViewModel<S : UiState, E : Any> : ViewModel() {
+abstract class BaseViewModel<S : UiState, E : Event> : ViewModel(), UiStateProvider {
 
     private val initialState: S by lazy { createInitialState() }
     abstract fun createInitialState(): S
@@ -19,10 +17,10 @@ abstract class BaseViewModel<S : UiState, E : Any> : ViewModel() {
         get() = uiState.value
 
     private val _uiState: MutableStateFlow<S> = MutableStateFlow(initialState)
-    val uiState = _uiState.asStateFlow()
+    override val uiState = _uiState.asStateFlow()
 
     private val _events: Channel<E> = Channel(Channel.UNLIMITED)
-    val events = _events.receiveAsFlow()
+    override val events = _events.receiveAsFlow()
 
     private val stateMutex = Mutex()
 
@@ -39,8 +37,14 @@ abstract class BaseViewModel<S : UiState, E : Any> : ViewModel() {
     }
 
     protected fun sendEvent(event: E) {
-        _events.trySend(event)
+        _events.offer(event)
     }
 }
 
 interface UiState
+interface Event
+
+interface UiStateProvider {
+    val uiState: StateFlow<UiState>
+    val events: Flow<Event>
+}
