--- app/src/main/java/com/tinkooladik/tvshows/base/UiError.kt
@@ -1,6 +1,0 @@
-package com.tinkooladik.tvshows.base
-
-sealed class UiError
-
-object NoConnection : UiError()
-object NoData : UiError()
