--- app/src/main/java/com/tinkooladik/tvshows/ui/NavActivity.kt
@@ -1,41 +1,0 @@
-package com.tinkooladik.tvshows.ui
-
-import android.os.Bundle
-import androidx.activity.compose.setContent
-import androidx.appcompat.app.AppCompatActivity
-import androidx.compose.material.Text
-import androidx.core.view.WindowCompat
-import androidx.navigation.NavController
-import androidx.navigation.fragment.NavHostFragment
-import com.google.accompanist.insets.ProvideWindowInsets
-import com.tinkooladik.tvshows.R
-import dagger.hilt.android.AndroidEntryPoint
-
-@AndroidEntryPoint
-class NavActivity : AppCompatActivity() {
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        //allow handling insets
-        WindowCompat.setDecorFitsSystemWindows(window, false)
-
-        setContent {
-            ProvideWindowInsets(consumeWindowInsets = false) {
-                Text(text = "Nav Activity")
-            }
-        }
-    }
-
-    override fun onSupportNavigateUp(): Boolean {
-        return findNavController().navigateUp() || super.onSupportNavigateUp()
-    }
-
-    /**
-     * See https://issuetracker.google.com/142847973
-     */
-    private fun findNavController(): NavController {
-        val navHostFragment =
-            supportFragmentManager.findFragmentById(R.id.nav_host_fragment) as NavHostFragment
-        return navHostFragment.navController
-    }
-}
