--- app/src/main/java/com/tinkooladik/tvshows/ui/home/HomeFragment.kt
@@ -1,40 +1,0 @@
-package com.tinkooladik.tvshows.ui.home
-
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import androidx.compose.material.Text
-import androidx.compose.runtime.CompositionLocalProvider
-import androidx.compose.ui.platform.ComposeView
-import androidx.fragment.app.Fragment
-import androidx.fragment.app.viewModels
-import com.google.accompanist.insets.LocalWindowInsets
-import com.google.accompanist.insets.ViewWindowInsetObserver
-import com.tinkooladik.tvshows.ui.theme.TvShowsTheme
-
-class HomeFragment : Fragment() {
-
-    private val viewModel: HomeViewModel by viewModels()
-
-    override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View = ComposeView(inflater.context).apply {
-        layoutParams = ViewGroup.LayoutParams(
-            ViewGroup.LayoutParams.MATCH_PARENT,
-            ViewGroup.LayoutParams.MATCH_PARENT
-        )
-
-        val windowInsets = ViewWindowInsetObserver(this).start()
-
-        setContent {
-            CompositionLocalProvider(LocalWindowInsets provides windowInsets) {
-                TvShowsTheme {
-                    Text(text = "Home Fragment")
-                }
-            }
-        }
-    }
-
-}
