+++ build.gradle
@@ -29,7 +29,7 @@ buildscript {
     }
     dependencies {
         classpath "com.android.tools.build:gradle:7.0.0"
-        classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:1.5.10"
+        classpath 'org.jetbrains.kotlin:kotlin-gradle-plugin:1.5.10'
         classpath 'com.google.dagger:hilt-android-gradle-plugin:2.38.1'
     }
 }
