+++ common/build.gradle
@@ -1,0 +1,57 @@
+plugins {
+    id "com.android.library"
+    id 'kotlin-android'
+    id 'kotlin-kapt'
+    id 'dagger.hilt.android.plugin'
+}
+
+android {
+    compileSdk buildConfig.compileSdk
+
+    compileOptions {
+        sourceCompatibility JavaVersion.VERSION_1_8
+        targetCompatibility JavaVersion.VERSION_1_8
+    }
+    kotlinOptions {
+        jvmTarget = '1.8'
+    }
+    kapt {
+        javacOptions {
+            // These options are normally set automatically via the Hilt Gradle plugin, but we
+            // set them manually to workaround a bug in the Kotlin 1.5.20
+            option("-Adagger.fastInit=ENABLED")
+            option("-Adagger.hilt.android.internal.disableAndroidSuperclassValidation=true")
+        }
+    }
+}
+
+dependencies {
+    api project(":domain")
+    api project(":data")
+
+    api "androidx.core:core-ktx:$version_ktx"
+    api "androidx.appcompat:appcompat:$version_appcompat"
+    api "com.google.android.material:material:$version_material"
+
+    api "androidx.lifecycle:lifecycle-runtime-ktx:$version_lifecycle"
+    api "androidx.lifecycle:lifecycle-viewmodel-ktx:$version_lifecycle"
+
+    api "androidx.fragment:fragment-ktx:$version_fragment"
+    api "androidx.navigation:navigation-fragment-ktx:$version_navigation"
+    api "androidx.navigation:navigation-ui-ktx:$version_navigation"
+
+    api "org.jetbrains.kotlinx:kotlinx-coroutines-core:$version_coroutines"
+
+    api "com.michael-bull.kotlin-result:kotlin-result:$version_result"
+    api "com.michael-bull.kotlin-result:kotlin-result-coroutines:$version_result"
+
+    implementation "com.google.dagger:hilt-android:$version_hilt"
+    kapt "com.google.dagger:hilt-android-compiler:$version_hilt"
+
+    api "com.jakewharton.timber:timber:$version_timber"
+
+    testApi "junit:junit:$version_junit"
+    androidTestApi 'androidx.test.ext:junit:1.1.3'
+    androidTestApi "androidx.test.espresso:espresso-core:$version_espresso"
+    androidTestApi "androidx.compose.ui:ui-test-junit4:$version_compose"
+}
