+++ common/src/main/java/com/tinkooladik/tvshows/common/UiError.kt
@@ -1,0 +1,6 @@
+package com.tinkooladik.tvshows.common
+
+sealed class UiError
+
+object NoConnection : UiError()
+object NoData : UiError()
