+++ common/src/main/res/mipmap-mdpi/ic_launcher.webp
